Updated notification flow to get rid of unnecessary read/write actions to Firebase. Added hooks folder to help handle notification related functionality.
